+++ inst/include/dust/dust.hpp
@@ -169,9 +169,12 @@ public:
     size_t blockSize = 128;
     size_t blockCount;
     bool use_shared_L1 = true;
+    size_t n_shared_int_effective = _device_data.n_shared_int +
+      align_padding(_device_data.n_shared_int * sizeof(int), sizeof(real_t))
+      / sizeof(int);
     size_t shared_size_bytes =
-      _device_data.n_shared_int * n_pars_effective() * sizeof(int) +
-      _device_data.n_shared_real * n_pars_effective() * sizeof(real_t);
+      (n_shared_int_effective * sizeof(int) +
+       _device_data.n_shared_real * sizeof(real_t);
     if (_n_particles_each < warp_size || shared_size_bytes > _shared_size) {
       // If not enough particles per pars to make a whole block use
       // shared, or if shared_t too big for L1, turn it off, and run