+++ NEWS.md
@@ -1,6 +1,7 @@
 # dust 0.8.3
 
 * Fix issue with `rnorm()` running on a GPU (device code).
+* Fix issue with unaligned shared copy in CUDA code.
 
 # dust 0.8.2